// assets/games/tongits/script/proto/tongits.ts
export interface TongitsPlayerInfo {
  /** 基础玩家信息 */
  playerInfo:
    | PlayerInfo
    | undefined;
  /** 手牌数量 (对其他玩家可见) */
  handCardCount: number;
  /** 是否是庄家 */
  isDealer: boolean;
  /** 已经打出的牌组 (明牌) */
  displayedMelds: Meld[];
  /** 自己的手牌 老 */
  handCards: number[];
  /** 是否可以发起挑战 */
  isFight: boolean;
  /** 时间戳 */
  countdown: number;
  /** 1:默认待选择 2:发起 3:接受 4:拒绝 5:烧死 */
  changeStatus: number;
  /** 1:不可操作, 2:select, 3:Action */
  status: number;
  /** 是否赢得比赛 */
  isWin: boolean;
  /** 点数 */
  cardPoint: number;
  /** 自动组牌 */
  isAuto: boolean;
  /** 自己的手牌 新 */
  groupCards: Cards[];
}
// ...
export interface GameInfo {
  /** 当前操作玩家ID */
  actionPlayerId: number;
  /** 游戏阶段 1:初始 2:游戏中  3: 挑战中 4: 结算前  5:结算中 */
  status: number;
  /** 牌堆剩余牌数 */
  deckCardCount: number;
  /** 弃牌堆的牌 从下到上 */
  discardPile: number[];
  /** 底池信息 */
  pot:
    | PotInfo
    | undefined;
  /** 弃牌显示牌 */
  discardCard: number;
  /** 胜利类型  1 tongits 2 挑战 3 时间结束比大小: */
  winType: number;
  /** 视角玩家id */
  perspectiveId: number;
  /** 当前下注币值 */
  betAmount: number;
  /** 币值列表 */
  scores: number[];
}

/** 底池信息 */
export interface PotInfo {
  /** 底池金额 */
  base: number;
  /** 连续未被赢取的次数 */
  winCount: number;
  /** 当前所属玩家id */
  useId: number;
}

export interface PlayerResult {
  playerInfo:
    | TongitsPlayerInfo
    | undefined;
  /** 总输赢获得的钱 */
  sumWinBonus: number;
  /** 正常输赢获得的钱 */
  normalWinBonus: number;
  /** Tongits输赢获得的钱 */
  tongitsWinBonus: number;
  /** 牌型奖励获得的钱 */
  cardTypeBonus: number;
  /** 奖励点 宝牌获得的钱 */
  bonusBonus: number;
  /** 烧牌获得的钱 */
  burnedBonus: number;
  /** 赢得挑战获得的钱 */
  winChallengeBonus: number;
  /** 底池获得的钱 */
  potBonus: number;
  /** 点数 */
  cardPoint: number;
}

/**
 * 进入房间
 * MessageType: TONGITS_JOIN_ROOM_REQ (3001)
 */
export interface JoinRoomReq {
  roomId: number;
}

/**
 * 进入房间
 * MessageType: TONGITS_JOIN_ROOM_RES (3002)
 */
export interface JoinRoomRes {
  /** 房间信息 */
  roomInfo: RoomInfo | undefined;
  players: TongitsPlayerInfo[];
  /** 旁观者信息 */
  watchers: PlayerInfo[];
  /** 旁观者数量 */
  playersCount: number;
  /** 开麦用户列表 */
  speakers: PlayerInfo[];
  /** 自己信息 */
  self:
    | TongitsPlayerInfo
    | undefined;
  /** 游戏信息 */
  gameInfo:
    | GameInfo
    | undefined;
  /** 玩家设置 */
  playerSettings: PlayerSettings | undefined;
}

/**
 * 开始游戏
 * MessageType: TONGITS_ROOM_OWNER_START_GAME_REQ (3019)
 */
export interface StartGameReq {
}

/**
 * 开始游戏
 * MessageType: TONGITS_ROOM_OWNER_START_GAME_RES (3020)
 */
export interface StartGameRes {
}

/**
 * 游戏开始广播
 * MessageType: TONGITS_START_GAME_BROADCAST (3015)
 */
export interface GameStartBroadcast {
  /** 游戏初始状态 */
  gameInfo:
    | GameInfo
    | undefined;
  /** 座位上的玩家信息 */
  players: TongitsPlayerInfo[];
  userId: number;
}
// ...
export interface ActionChangeBroadcast {
  /** 操作玩家 */
  actionPlayerId: number;
  /** 倒计时 */
  countdown: number;
  /** 是否可以发起挑战 */
  isFight: boolean;
  /** 1:不可操作, 2:select, 3:Action */
  status: number;
  /** 自己的手牌 */
  groupCards: Cards[];
  /** 通知玩家id */
  userId: number;
}
// ...
export const PotInfo = {
  encode(message: PotInfo, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.base !== 0) {
      writer.uint32(8).int32(message.base);
    }
    if (message.winCount !== 0) {
      writer.uint32(16).int32(message.winCount);
    }
    if (message.useId !== 0) {
      writer.uint32(24).int64(message.useId);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): PotInfo {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePotInfo();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.base = reader.int32();
          continue;
        case 2:
          if (tag !== 16) {
            break;
          }

          message.winCount = reader.int32();
          continue;
        case 3:
          if (tag !== 24) {
            break;
          }

          message.useId = longToNumber(reader.int64() as Long);
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PotInfo {
    return {
      base: isSet(object.base) ? globalThis.Number(object.base) : 0,
      winCount: isSet(object.winCount) ? globalThis.Number(object.winCount) : 0,
      useId: isSet(object.useId) ? globalThis.Number(object.useId) : 0,
    };
  },

  toJSON(message: PotInfo): unknown {
    const obj: any = {};
    if (message.base !== 0) {
      obj.base = Math.round(message.base);
    }
    if (message.winCount !== 0) {
      obj.winCount = Math.round(message.winCount);
    }
    if (message.useId !== 0) {
      obj.useId = Math.round(message.useId);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<PotInfo>, I>>(base?: I): PotInfo {
    return PotInfo.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PotInfo>, I>>(object: I): PotInfo {
    const message = createBasePotInfo();
    message.base = object.base ?? 0;
    message.winCount = object.winCount ?? 0;
    message.useId = object.useId ?? 0;
    return message;
  },
};
// ...
export const GameStartBroadcast = {
  encode(message: GameStartBroadcast, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.gameInfo !== undefined) {
      GameInfo.encode(message.gameInfo, writer.uint32(10).fork()).ldelim();
    }
    for (const v of message.players) {
      TongitsPlayerInfo.encode(v!, writer.uint32(18).fork()).ldelim();
    }
    if (message.userId !== 0) {
      writer.uint32(32).int64(message.userId);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): GameStartBroadcast {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGameStartBroadcast();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.gameInfo = GameInfo.decode(reader, reader.uint32());
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.players.push(TongitsPlayerInfo.decode(reader, reader.uint32()));
          continue;
        case 4:
          if (tag !== 32) {
            break;
          }

          message.userId = longToNumber(reader.int64() as Long);
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GameStartBroadcast {
    return {
      gameInfo: isSet(object.gameInfo) ? GameInfo.fromJSON(object.gameInfo) : undefined,
      players: globalThis.Array.isArray(object?.players)
        ? object.players.map((e: any) => TongitsPlayerInfo.fromJSON(e))
        : [],
      userId: isSet(object.userId) ? globalThis.Number(object.userId) : 0,
    };
  },

  toJSON(message: GameStartBroadcast): unknown {
    const obj: any = {};
    if (message.gameInfo !== undefined) {
      obj.gameInfo = GameInfo.toJSON(message.gameInfo);
    }
    if (message.players?.length) {
      obj.players = message.players.map((e) => TongitsPlayerInfo.toJSON(e));
    }
    if (message.userId !== 0) {
      obj.userId = Math.round(message.userId);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GameStartBroadcast>, I>>(base?: I): GameStartBroadcast {
    return GameStartBroadcast.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GameStartBroadcast>, I>>(object: I): GameStartBroadcast {
    const message = createBaseGameStartBroadcast();
    message.gameInfo = (object.gameInfo !== undefined && object.gameInfo !== null)
      ? GameInfo.fromPartial(object.gameInfo)
      : undefined;
    message.players = object.players?.map((e) => TongitsPlayerInfo.fromPartial(e)) || [];
    message.userId = object.userId ?? 0;
    return message;
  },
};
// ...
export const ActionChangeBroadcast = {
  encode(message: ActionChangeBroadcast, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.actionPlayerId !== 0) {
      writer.uint32(8).int64(message.actionPlayerId);
    }
    if (message.countdown !== 0) {
      writer.uint32(16).int32(message.countdown);
    }
    if (message.isFight !== false) {
      writer.uint32(24).bool(message.isFight);
    }
    if (message.status !== 0) {
      writer.uint32(32).int32(message.status);
    }
    for (const v of message.groupCards) {
      Cards.encode(v!, writer.uint32(42).fork()).ldelim();
    }
    if (message.userId !== 0) {
      writer.uint32(88).int64(message.userId);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): ActionChangeBroadcast {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseActionChangeBroadcast();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.actionPlayerId = longToNumber(reader.int64() as Long);
          continue;
        case 2:
          if (tag !== 16) {
            break;
          }

          message.countdown = reader.int32();
          continue;
        case 3:
          if (tag !== 24) {
            break;
          }

          message.isFight = reader.bool();
          continue;
        case 4:
          if (tag !== 32) {
            break;
          }

          message.status = reader.int32();
          continue;
        case 5:
          if (tag !== 42) {
            break;
          }

          message.groupCards.push(Cards.decode(reader, reader.uint32()));
          continue;
        case 11:
          if (tag !== 88) {
            break;
          }

          message.userId = longToNumber(reader.int64() as Long);
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ActionChangeBroadcast {
    return {
      actionPlayerId: isSet(object.actionPlayerId) ? globalThis.Number(object.actionPlayerId) : 0,
      countdown: isSet(object.countdown) ? globalThis.Number(object.countdown) : 0,
      isFight: isSet(object.isFight) ? globalThis.Boolean(object.isFight) : false,
      status: isSet(object.status) ? globalThis.Number(object.status) : 0,
      groupCards: globalThis.Array.isArray(object?.groupCards)
        ? object.groupCards.map((e: any) => Cards.fromJSON(e))
        : [],
      userId: isSet(object.userId) ? globalThis.Number(object.userId) : 0,
    };
  },

  toJSON(message: ActionChangeBroadcast): unknown {
    const obj: any = {};
    if (message.actionPlayerId !== 0) {
      obj.actionPlayerId = Math.round(message.actionPlayerId);
    }
    if (message.countdown !== 0) {
      obj.countdown = Math.round(message.countdown);
    }
    if (message.isFight !== false) {
      obj.isFight = message.isFight;
    }
    if (message.status !== 0) {
      obj.status = Math.round(message.status);
    }
    if (message.groupCards?.length) {
      obj.groupCards = message.groupCards.map((e) => Cards.toJSON(e));
    }
    if (message.userId !== 0) {
      obj.userId = Math.round(message.userId);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ActionChangeBroadcast>, I>>(base?: I): ActionChangeBroadcast {
    return ActionChangeBroadcast.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ActionChangeBroadcast>, I>>(object: I): ActionChangeBroadcast {
    const message = createBaseActionChangeBroadcast();
    message.actionPlayerId = object.actionPlayerId ?? 0;
    message.countdown = object.countdown ?? 0;
    message.isFight = object.isFight ?? false;
    message.status = object.status ?? 0;
    message.groupCards = object.groupCards?.map((e) => Cards.fromPartial(e)) || [];
    message.userId = object.userId ?? 0;
    return message;
  },
};